+++ POSserver/POSserver/urls.py
@@ -13,12 +13,13 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.conf.urls import url
+# from django.conf.urls import url
 from django.contrib import admin
 from graphene_django.views import GraphQLView
 from django.views.decorators.csrf import csrf_exempt
 from django.urls import path, include, re_path
 from .views import GeneratePDF
+from payment.views import checkout
 
 
 urlpatterns = [
@@ -26,4 +27,6 @@ urlpatterns = [
     path("accounts/", include("allauth.urls")),
     re_path(r"^graphql/$", csrf_exempt(GraphQLView.as_view(graphiql=True))),
     re_path(r"^pdf/$", csrf_exempt(GeneratePDF.as_view())),
+    path(r'create-charge/', checkout, name="checkout"),
+    
 ]
