+++ POSserver/payment/urls.py
@@ -1,0 +1,6 @@
+from .views import checkout
+from django.urls import path
+
+urlpatterns = [
+  path(r'create-charge/', checkout),
+]
