+++ client/src/components/index.js
@@ -3,7 +3,7 @@ import AddTag from "./tags/addtag";
 import TagView from "./tags/tagview";
 import Settings from "./settings/settings";
 import Billing from "./billing/billing";
-import CheckoutForm from "./billing/checkoutform"
+import Checkout from "./billing/checkout"
 import Parts from "./parts/parts";
 import AddPart from "./parts/addpart";
 import PartView from "./parts/partview";
